requirements.txt
flask              
Flask-SQLAlchemy   
greenlet           
idna               
imbalanced-learn             
joblib         
numpy              
pandas
scikit-learn            
sklearn-compat
psycopg2

